requirements.txt
# Core AI and ML dependencies
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.39.0

# Audio processing
openai-whisper>=20230314
soundfile>=0.12.1
librosa>=0.10.0

# Vector search and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Database and data handling
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.12.0

# Web interface
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfplumber>=0.9.0
pypdf>=3.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
psutil>=5.9.0
